[refactor] Create a file for the ClassChecker in pylint.classes
